investment_portfolio/requirements.txt
django==4.2.5
djangorestframework==3.14.0
psycopg2==2.9.6
redis